batch/requirements.txt
pandas==2.1.4
numpy==1.26.4
psycopg2-binary==2.9.9
boto3==1.34.0
pyarrow==15.0.0
